fwd dec done

+++ RipTag/Source/Input/Input.cpp
@@ -557,8 +557,8 @@ void Input::ResetMouse()
 {
 	if (InputHandler::getWindowFocus())
 	{
-		int midX = InputHandler::getviewportPos().x + (InputHandler::getWindowSize().x / 2.0f);
-		int midY = InputHandler::getviewportPos().y + (InputHandler::getWindowSize().y / 2.0f);
+		int midX = InputHandler::getviewportPos().x + (int)(InputHandler::getWindowSize().x / 2.0f);
+		int midY = InputHandler::getviewportPos().y + (int)(InputHandler::getWindowSize().y / 2.0f);
 
 		SetCursorPos(midX, midY);
 	}
